moved app_host to managers
